Move CSV API endpoint into V2 fixes #3045
