+++ app/experimenter/experiments/api/v1/urls.py
@@ -1,7 +1,6 @@
 from django.conf.urls import url
 
 from experimenter.experiments.api.v1.views import (
-    ExperimentCSVListView,
     ExperimentDetailView,
     ExperimentListView,
     ExperimentRecipeView,
@@ -9,7 +8,6 @@ from experimenter.experiments.api.v1.views import (
 
 
 urlpatterns = [
-    url(r"^csv/$", ExperimentCSVListView.as_view(), name="experiments-api-csv",),
     url(
         r"^(?P<slug>[\w-]+)/recipe/$",
         ExperimentRecipeView.as_view(),
