+++ app/experimenter/experiments/tests/api/v1/test_views.py
@@ -4,18 +4,12 @@ from django.conf import settings
 from django.test import TestCase
 from django.urls import reverse
 from parameterized import parameterized
-from rest_framework_csv.renderers import CSVRenderer
 
 from experimenter.experiments.constants import ExperimentConstants
 from experimenter.experiments.models import Experiment
-from experimenter.experiments.api.v1.serializers import (
-    ExperimentSerializer,
-    ExperimentCSVSerializer,
-)
+from experimenter.experiments.api.v1.serializers import ExperimentSerializer
 from experimenter.experiments.tests.factories import ExperimentFactory
 from experimenter.normandy.serializers import ExperimentRecipeSerializer
-from experimenter.openidc.tests.factories import UserFactory
-from experimenter.projects.tests.factories import ProjectFactory
 
 
 class TestExperimentListView(TestCase):
@@ -119,78 +113,3 @@ class TestExperimentRecipeView(TestCase):
             **{settings.OPENIDC_EMAIL_HEADER: user_email},
         )
         self.assertEqual(response.status_code, 404)
-
-
-class TestExperimentCSVListView(TestCase):
-    def test_get_returns_csv_info(self):
-        user_email = "user@example.com"
-        experiment1 = ExperimentFactory.create_with_status(
-            Experiment.STATUS_DRAFT, name="a"
-        )
-        experiment2 = ExperimentFactory.create_with_status(
-            Experiment.STATUS_DRAFT, name="b"
-        )
-
-        response = self.client.get(
-            reverse("experiments-api-csv"), **{settings.OPENIDC_EMAIL_HEADER: user_email},
-        )
-
-        self.assertEqual(response.status_code, 200)
-
-        csv_data = response.content
-        expected_csv_data = CSVRenderer().render(
-            ExperimentCSVSerializer([experiment1, experiment2], many=True).data,
-            renderer_context={"header": ExperimentCSVSerializer.Meta.fields},
-        )
-        self.assertEqual(csv_data, expected_csv_data)
-
-    def test_view_filters_by_project(self):
-        user_email = "user@example.com"
-        project = ProjectFactory.create()
-        experiment1 = ExperimentFactory.create_with_status(
-            Experiment.STATUS_DRAFT, name="a", projects=[project]
-        )
-        experiment2 = ExperimentFactory.create_with_status(
-            Experiment.STATUS_DRAFT, name="b", projects=[project]
-        )
-        ExperimentFactory.create_with_variants()
-
-        url = reverse("experiments-api-csv")
-        response = self.client.get(
-            f"{url}?projects={project.id}", **{settings.OPENIDC_EMAIL_HEADER: user_email}
-        )
-
-        self.assertEqual(response.status_code, 200)
-
-        csv_data = response.content
-        expected_csv_data = CSVRenderer().render(
-            ExperimentCSVSerializer([experiment1, experiment2], many=True).data,
-            renderer_context={"header": ExperimentCSVSerializer.Meta.fields},
-        )
-        self.assertEqual(csv_data, expected_csv_data)
-
-    def test_view_filters_by_subscriber(self):
-        user = UserFactory(email="user@example.com")
-        experiment1 = ExperimentFactory.create_with_status(
-            Experiment.STATUS_DRAFT, name="a"
-        )
-        experiment1.subscribers.add(user)
-        experiment2 = ExperimentFactory.create_with_status(
-            Experiment.STATUS_DRAFT, name="b"
-        )
-        experiment2.subscribers.add(user)
-        ExperimentFactory.create_with_variants()
-
-        url = reverse("experiments-api-csv")
-        response = self.client.get(
-            f"{url}?subscribed=on", **{settings.OPENIDC_EMAIL_HEADER: user.email}
-        )
-
-        self.assertEqual(response.status_code, 200)
-
-        csv_data = response.content
-        expected_csv_data = CSVRenderer().render(
-            ExperimentCSVSerializer([experiment1, experiment2], many=True).data,
-            renderer_context={"header": ExperimentCSVSerializer.Meta.fields},
-        )
-        self.assertEqual(csv_data, expected_csv_data)
