+++ app/experimenter/experiments/tests/api/v2/test_views.py
@@ -4,22 +4,27 @@ from django.conf import settings
 from django.core import mail
 from django.test import TestCase
 from django.urls import reverse
+from rest_framework_csv.renderers import CSVRenderer
+
 
 from experimenter.experiments.constants import ExperimentConstants
 from experimenter.experiments.models import Experiment
 from experimenter.experiments.api.v2.serializers import (
-    ExperimentTimelinePopSerializer,
+    ExperimentCSVSerializer,
     ExperimentDesignAddonSerializer,
     ExperimentDesignGenericSerializer,
     ExperimentDesignMessageSerializer,
     ExperimentDesignMultiPrefSerializer,
     ExperimentDesignPrefSerializer,
+    ExperimentTimelinePopSerializer,
 )
 from experimenter.experiments.tests.factories import (
     ExperimentFactory,
     ExperimentVariantFactory,
     VariantPreferencesFactory,
 )
+from experimenter.openidc.tests.factories import UserFactory
+from experimenter.projects.tests.factories import ProjectFactory
 
 
 class TestExperimentSendIntentToShipEmailView(TestCase):
@@ -605,3 +610,78 @@ class TestExperimentTimelinePopulationView(TestCase):
         )
 
         self.assertEqual(response.status_code, 200)
+
+
+class TestExperimentCSVListView(TestCase):
+    def test_get_returns_csv_info(self):
+        user_email = "user@example.com"
+        experiment1 = ExperimentFactory.create_with_status(
+            Experiment.STATUS_DRAFT, name="a"
+        )
+        experiment2 = ExperimentFactory.create_with_status(
+            Experiment.STATUS_DRAFT, name="b"
+        )
+
+        response = self.client.get(
+            reverse("experiments-api-csv"), **{settings.OPENIDC_EMAIL_HEADER: user_email},
+        )
+
+        self.assertEqual(response.status_code, 200)
+
+        csv_data = response.content
+        expected_csv_data = CSVRenderer().render(
+            ExperimentCSVSerializer([experiment1, experiment2], many=True).data,
+            renderer_context={"header": ExperimentCSVSerializer.Meta.fields},
+        )
+        self.assertEqual(csv_data, expected_csv_data)
+
+    def test_view_filters_by_project(self):
+        user_email = "user@example.com"
+        project = ProjectFactory.create()
+        experiment1 = ExperimentFactory.create_with_status(
+            Experiment.STATUS_DRAFT, name="a", projects=[project]
+        )
+        experiment2 = ExperimentFactory.create_with_status(
+            Experiment.STATUS_DRAFT, name="b", projects=[project]
+        )
+        ExperimentFactory.create_with_variants()
+
+        url = reverse("experiments-api-csv")
+        response = self.client.get(
+            f"{url}?projects={project.id}", **{settings.OPENIDC_EMAIL_HEADER: user_email}
+        )
+
+        self.assertEqual(response.status_code, 200)
+
+        csv_data = response.content
+        expected_csv_data = CSVRenderer().render(
+            ExperimentCSVSerializer([experiment1, experiment2], many=True).data,
+            renderer_context={"header": ExperimentCSVSerializer.Meta.fields},
+        )
+        self.assertEqual(csv_data, expected_csv_data)
+
+    def test_view_filters_by_subscriber(self):
+        user = UserFactory(email="user@example.com")
+        experiment1 = ExperimentFactory.create_with_status(
+            Experiment.STATUS_DRAFT, name="a"
+        )
+        experiment1.subscribers.add(user)
+        experiment2 = ExperimentFactory.create_with_status(
+            Experiment.STATUS_DRAFT, name="b"
+        )
+        experiment2.subscribers.add(user)
+        ExperimentFactory.create_with_variants()
+
+        url = reverse("experiments-api-csv")
+        response = self.client.get(
+            f"{url}?subscribed=on", **{settings.OPENIDC_EMAIL_HEADER: user.email}
+        )
+
+        self.assertEqual(response.status_code, 200)
+
+        csv_data = response.content
+        expected_csv_data = CSVRenderer().render(
+            ExperimentCSVSerializer([experiment1, experiment2], many=True).data,
+            renderer_context={"header": ExperimentCSVSerializer.Meta.fields},
+        )
+        self.assertEqual(csv_data, expected_csv_data)
